tui/requirements.txt

textual>=0.44.0
rich>=13.7.0
pyyaml>=6.0.2
python-dotenv>=1.0.1
